requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
openai>=1.50.0
pydantic>=2.0.0
typing-extensions>=4.0.0
matplotlib>=3.0.0
networkx>=3.0
gradio>=6.0.0
gitpython>=3.1.0
